fix: circular dependency in dicovery (selectDeviceThunk) by sepearating the think into solo file
