+++ suite-common/wallet-core/src/device/deviceThunks.ts
@@ -44,46 +44,7 @@ import {
 } from './deviceSelectors';
 import { selectAccountByKey } from '../accounts/accountsSelectors';
 import { startDiscoveryThunk } from '../discovery/discoveryThunks';
-
-type SelectDeviceThunkParams = {
-    device: Device | TrezorDevice | undefined;
-};
-
-/**
- * Called from:
- * - `@trezor/connect` events handler `handleDeviceConnect`, `handleDeviceDisconnect`
- * - from user action in `@suite-components/DeviceMenu`
- */
-export const selectDeviceThunk = createThunk<void, SelectDeviceThunkParams, void>(
-    `${DEVICE_MODULE_PREFIX}/selectDevice`,
-    ({ device }, { dispatch, getState, extra }) => {
-        let trezorDevice: TrezorDevice | typeof undefined;
-        const devices = selectDevices(getState());
-
-        if (device) {
-            // "ts" is one of the field which distinguish Device from TrezorDevice
-            // (device from connect doesn't have timestamp but suite device has)
-            if ('ts' in device) {
-                // requested device is a @suite TrezorDevice type. get exact instance from reducer
-                trezorDevice = getSelectedDevice(device, devices);
-            } else {
-                // requested device is a @trezor/connect Device type
-                // find all instances and select recently used
-                const instances = devices.filter(d => d.path === device.path);
-
-                trezorDevice = sortByTimestamp(instances)[0];
-            }
-        }
-
-        dispatch(deviceActions.selectDevice(trezorDevice));
-        if (
-            trezorDevice?.state?.staticSessionId !== undefined &&
-            extra.selectors.selectSuiteSettings(getState()).isLocalFirstStorageEnabled
-        ) {
-            dispatch(extra.thunks.subscribeLocalFirstStorage({ device: trezorDevice }));
-        }
-    },
-);
+import { selectDeviceThunk } from '../discovery/selectDeviceThunk';
 
 /**
  * Toggles remembering the given device. I.e. if given device is not remembered it will become remembered
